BUG Refactor to better organization - avoid circular imports. Still need to refactor io.config. Rename AnalysisConfig to DescribedAnalysis
